# openapi/openapi.yaml
openapi: 3.0.0
paths:
...
  /tags:
    post:
      operationId: TagsController_create
      summary: 创建标签
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTagDto'
      responses:
        '201':
          description: 创建成功
        '409':
          description: 同名标签已存在
      tags: &ref_13
        - tags
      security: &ref_14
        - bearer: []
    get:
      operationId: TagsController_findAll
      summary: 获取标签列表
      parameters:
        - name: workspaceId
          required: false
          in: query
          description: 工作空间 ID（必填）
          schema:
            type: string
        - name: page
          required: false
          in: query
          description: 页码
          schema:
            default: 1
            type: number
        - name: pageSize
          required: false
          in: query
          description: 每页数量
          schema:
            default: 20
            type: number
      responses:
        '200':
          description: 获取成功
      tags: *ref_13
      security: *ref_14
  /tags/{tagId}/usage:
    get:
      operationId: TagsController_getUsage
      summary: 获取标签使用统计
      parameters:
        - name: tagId
          required: true
          in: path
          description: 标签 ID
          schema:
            type: string
      responses:
        '200':
          description: 返回 usage（使用该标签的文档数）
      tags: *ref_13
      security: *ref_14
  /tags/{tagId}:
    get:
      operationId: TagsController_findOne
      summary: 获取标签详情
      parameters:
        - name: tagId
          required: true
          in: path
          description: 标签 ID
          schema:
            type: string
      responses:
        '200':
          description: 获取成功
        '404':
          description: 标签不存在
      tags: *ref_13
      security: *ref_14
    patch:
      operationId: TagsController_update
      summary: 更新标签
      parameters:
        - name: tagId
          required: true
          in: path
          description: 标签 ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTagDto'
      responses:
        '200':
          description: 更新成功
        '404':
          description: 标签不存在
        '409':
          description: 同名标签已存在
      tags: *ref_13
      security: *ref_14
    delete:
      operationId: TagsController_remove
      summary: 删除标签
      parameters:
        - name: tagId
          required: true
          in: path
          description: 标签 ID
          schema:
            type: string
      responses:
        '200':
          description: 删除成功，并从所有文档的 tags 中移除
        '404':
          description: 标签不存在
      tags: *ref_13
      security: *ref_14
...
  /search:
    get:
      operationId: SearchController_globalSearch
      summary: 全局搜索
      parameters:
        - name: query
          required: true
          in: query
          description: 搜索关键词
          schema:
            example: 文档标题或块内容
            type: string
        - name: workspaceId
          required: false
          in: query
          description: 工作空间ID（限制搜索范围）
          schema:
            example: ws_1234567890_abc123
            type: string
        - name: type
          required: false
          in: query
          description: 搜索类型：doc=仅文档, block=仅块内容, all=全部
          schema:
            default: all
            enum:
              - doc
              - block
              - all
            type: string
        - name: page
          required: false
          in: query
          description: 页码
          schema:
            default: 1
            example: 1
            type: number
        - name: pageSize
          required: false
          in: query
          description: 每页数量
          schema:
            default: 20
            example: 20
            type: number
      responses:
        '200':
          description: 返回文档与块匹配结果
      tags: &ref_19
        - 搜索
      security: &ref_20
        - bearer: []
  /search/advanced:
    post:
      operationId: SearchController_advancedSearch
      summary: 高级搜索
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdvancedSearchDto'
      responses:
        '200':
          description: 支持标签、时间范围、创建者、排序
      tags: *ref_19
      security: *ref_20
...
info:
  title: 知识库 API
  description: 个人知识库系统 API 文档
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateTagDto:
      type: object
      properties:
        workspaceId:
          type: string
          description: 工作空间ID
          example: ws_1234567890_abc123
        name:
          type: string
          description: 标签名称
          example: 重要
          maxLength: 50
        color:
          type: string
          description: 标签颜色（十六进制）
          example: '#ff4d4f'
          maxLength: 20
      required:
        - workspaceId
        - name
    UpdateTagDto:
      type: object
      properties:
        name:
          type: string
          description: 标签名称
          maxLength: 50
        color:
          type: string
          description: 标签颜色
          maxLength: 20
...
    AdvancedSearchDto:
      type: object
      properties:
        query:
          type: string
          description: 搜索关键词
          example: 关键词
        workspaceId:
          type: string
          description: 工作空间ID
        type:
          type: string
          description: 搜索类型
          enum:
            - doc
            - block
            - all
        tags:
          description: 标签筛选
          example:
            - 标签1
          type: array
          items:
            type: string
        startDate:
          type: string
          description: 开始日期（ISO）
          example: '2026-01-01'
        endDate:
          type: string
          description: 结束日期（ISO）
          example: '2026-12-31'
        createdBy:
          type: string
          description: 创建者 userId
        sortBy:
          type: string
          description: 排序字段
          enum:
            - rank
            - updatedAt
            - createdAt
          default: rank
        sortOrder:
          type: string
          description: 排序方向
          enum:
            - asc
            - desc
          default: desc
        page:
          type: number
          description: 页码
          default: 1
        pageSize:
          type: number
          description: 每页数量
          default: 20
      required:
        - query
...